+++ slash/core/error.py
@@ -28,6 +28,10 @@ class Error(object):
             self.traceback = distill_traceback(tb, frame_correction=frame_correction)
         else:
             self.traceback = distill_call_stack(frame_correction=frame_correction+4)
+        self._is_failure = False
+
+    def mark_as_failure(self):
+        self._is_failure = True
 
     def is_fatal(self):
         if self._fatal:
@@ -39,7 +43,7 @@ class Error(object):
         return self
 
     def is_failure(self):
-        return isinstance(self.exception, FAILURE_EXCEPTION_TYPES)
+        return self._is_failure or isinstance(self.exception, FAILURE_EXCEPTION_TYPES)
 
     @classmethod
     def capture_exception(cls, exc_info=None):