+++ slash/plugins/builtin/xunit.py
@@ -48,13 +48,13 @@ class Plugin(PluginInterface):
 
     def error_added(self, result, error): # pylint: disable=unused-argument
         if error.is_failure():
-            self._add_error("failure")
+            self._add_error("failure", error)
         else:
-            self._add_error("error")
+            self._add_error("error", error)
 
-    def _add_error(self, errortype):
+    def _add_error(self, errortype, error):
         exc_type, exc_value, _ = exc_info = sys.exc_info()
-        self._add_element(errortype, {'type': exc_type.__name__ if exc_type else '', 'message': str(exc_value)}, text=get_traceback_string(exc_info) if exc_value is not None else None)
+        self._add_element(errortype, {'type': exc_type.__name__ if exc_type else errortype, 'message': error.message}, text=get_traceback_string(exc_info) if exc_value is not None else None)
 
     def _add_element(self, tag, attrib, text=None):
         if not context.test: