+++ tests/test_requirements.py
@@ -15,9 +15,16 @@ def test_requirements_mismatch_session_success(suite, suite_test):
 @pytest.mark.parametrize('requirement_fullfilled', [True, False])
 @pytest.mark.parametrize('use_message', [True, False])
 @pytest.mark.parametrize('use_fixtures', [True, False])
-def test_requirements(suite, suite_test, requirement_fullfilled, use_fixtures, use_message):
+@pytest.mark.parametrize('message_in_retval', [True, False])
+def test_requirements(suite, suite_test, requirement_fullfilled, use_fixtures, use_message, message_in_retval):
 
-    suite_test.add_decorator('slash.requires(lambda: {0})'.format(requirement_fullfilled))
+    message = "requires something very important"
+    if use_message and message_in_retval:
+        retval = '({}, {!r})'.format(requirement_fullfilled, message)
+    else:
+        retval = requirement_fullfilled
+
+    suite_test.add_decorator('slash.requires((lambda: {0}), {1!r})'.format(retval, message if use_message and not message_in_retval else ''))
     if not requirement_fullfilled:
         suite_test.expect_skip()
 
@@ -30,6 +37,9 @@ def test_requirements(suite, suite_test, requirement_fullfilled, use_fixtures, u
     else:
         assert not results[suite_test].is_started()
         assert results[suite_test].is_skip()
+        if use_message:
+            [skip] = results[suite_test].get_skips()
+            assert message in skip
 
 
 def test_requirements_on_class():